CBD-6006, changes to support ubuntu24 on azure and gcp in 8.0.0
Update couchbase-server.service with delegation settings.
This is a subset of https://review.couchbase.org/c/cloud-build-tools/+/218158, which is not ready to be merged due to perf issue on AWS.

I have received requests to patch gcp and azure first since the images are currently broken.  This change looks for create-provisioned-cgroups.sh to determine whether to update the service file.  Hence, this won't impact the images of couchbase-server 7.x and couchbase-columnar 1.x

-Ming

Change-Id: I8d4bf298a3e6dffc2b1a1a1d0fd06f2e7bf9bafc
Reviewed-on: https://review.couchbase.org/c/cloud-build-tools/+/230153
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Ming Ho <ming.ho@couchbase.com>