AV-115245, revamped datastore-agent AMI
Per discussion with Talina and Akhil, this one-off AMI is not
couchbase-server AMI + datastore.  Instead, it is: couchbase-server +
dp-runtime-agent + datastore-agent + dp-observer, etc.
Since it has no dependency on dp-agent and amzn2 will be EOL in 1/2 yr,
I am going to use ubuntu24 as the base image as other couchbase AMIs.

-Ming

Change-Id: Ie37b787bf4f0403b4e8040db7a5283c9369ca1a4
Reviewed-on: https://review.couchbase.org/c/cloud-build-tools/+/236115
Tested-by: Talina Shrotriya
Reviewed-by: Talina Shrotriya
Reviewed-by: Ravuri Sai Ram Akhil <akhil.ravuri@couchbase.com>